+++ engine/src/App.js
@@ -24,8 +24,6 @@ class App {
 		this._sizes = sizes;
 		this._isImageSmoothingEnabled = isImageSmoothingEnabled;
 
-		
-
 		/**
 		 * An array of global variables
 		 * @private