+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * On clicking on not object list - deselecting all objects.
 * Editor is not adding objects borders coordinates and their central points when saving.
 * In the objects settings you can now make them invisible.
-* Added text object type support.
+* Added text and sprite object types support.
 * Added dark theme modals and inputs.
 
 ### Engine changes:
@@ -16,6 +16,7 @@
 * Sprite setting `isRepeat` was renamed to `isRepeating`.
 * The default value for sprite setting `interval` was set to 60ms.
 * Was added a new sprite setting `isPlaying` for play or stop animation when necessary.
+* Improved API documentation.
 
 ## v0.1.2
 ### Editor changes: